Move the class to an appropriate location. Also, remove the overriding test methods.
